sbus: change domain's connection name
