+++ src/providers/data_provider_be.c
@@ -589,7 +589,7 @@ errno_t be_process_init(TALLOC_CTX *mem_ctx,
         goto done;
     }
 
-    be_ctx->sbus_name = sss_iface_domain_bus(be_ctx, be_ctx->domain);
+    be_ctx->sbus_name = confdb_get_domain_bus(be_ctx, be_ctx->domain);
     if (be_ctx->sbus_name == NULL) {
         DEBUG(SSSDBG_FATAL_FAILURE, "Could not get sbus backend name.\n");
         ret = ENOMEM;
