+++ src/responder/common/responder_common.c
@@ -752,7 +752,7 @@ sss_dp_init(struct resp_ctx *rctx,
         goto done;
     }
 
-    be_conn->bus_name = sss_iface_domain_bus(be_conn, domain);
+    be_conn->bus_name = confdb_get_domain_bus(be_conn, domain);
     if (be_conn->bus_name == NULL) {
         DEBUG(SSSDBG_FATAL_FAILURE, "Could not locate DP address.\n");
         ret = ENOMEM;
@@ -1522,7 +1522,7 @@ int sss_dp_get_domain_conn(struct resp_ctx *rctx, const char *domain,
     if (!rctx->be_conns) return ENOENT;
 
     for (iter = rctx->be_conns; iter; iter = iter->next) {
-        if (strcasecmp(domain, iter->domain->name) == 0) break;
+        if (strcasecmp(domain, iter->domain->conn_name) == 0) break;
     }
 
     if (!iter) return ENOENT;
